test(unit): move a11y basic test to utils
