+++ packages/vkui/src/testing/utils.tsx
@@ -1,11 +1,19 @@
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
 import { render, RenderResult, screen } from '@testing-library/react';
+import { configureAxe, toHaveNoViolations } from 'jest-axe';
 import { AdaptivityProps } from '../components/AdaptivityProvider/AdaptivityContext';
 import { AdaptivityProvider } from '../components/AdaptivityProvider/AdaptivityProvider';
 import { ScrollContext } from '../components/AppRoot/ScrollContext';
 import { ImgOnlyAttributes } from '../lib/utils';
 import { HasChildren } from '../types';
+
+export const axe = configureAxe({
+  /**
+   * @see https://github.com/dequelabs/axe-core/blob/develop/doc/rule-descriptions.md
+   */
+});
+expect.extend(toHaveNoViolations);
 
 export function fakeTimers() {
   beforeEach(() => jest.useFakeTimers());
@@ -41,6 +49,7 @@ export type ComponentTestOptions = {
   className?: boolean;
   style?: boolean;
   adaptivity?: AdaptivityProps;
+  a11y?: boolean;
 };
 
 type BasicProps = { style?: any; className?: string };
@@ -57,6 +66,15 @@ export function mountTest(Component: React.ComponentType<any>) {
   });
 }
 
+export function a11yTest(Component: React.ComponentType<any>) {
+  it('a11y: has no violations', async () => {
+    const { container } = render(<Component />);
+
+    const results = await axe(container);
+    expect(results).toHaveNoViolations();
+  });
+}
+
 export function baselineComponent<Props extends BasicProps>(
   RawComponent: React.ComponentType<Props>,
   {
@@ -64,6 +82,7 @@ export function baselineComponent<Props extends BasicProps>(
     style = true,
     className = true,
     domAttr = true,
+    a11y = true,
     adaptivity,
   }: ComponentTestOptions = {},
 ) {
@@ -75,6 +94,7 @@ export function baselineComponent<Props extends BasicProps>(
       )
     : RawComponent;
   mountTest(Component);
+  a11y && a11yTest(Component);
   forward &&
     it('forwards attributes', async () => {
       const cls = 'Custom';
